Extract hash→pin routing into useGlobeHashRoute hook
